+++ renderer/src/web/item-search/widget.ts
@@ -1,0 +1,6 @@
+import type { Widget, Anchor } from '../overlay/widgets.js'
+
+export interface ItemSearchWidget extends Widget {
+  anchor: Anchor
+  ocrGemsKey: string | null
+}
