+++ renderer/src/web/overlay/OverlayWindow.vue
@@ -42,7 +42,7 @@ import PriceCheckWindow from '@/web/price-check/PriceCheckWindow.vue'
 import WidgetItemCheck from '@/web/item-check/WidgetItemCheck.vue'
 import WidgetImageStrip from './WidgetImageStrip.vue'
 import WidgetDelveGrid from './WidgetDelveGrid.vue'
-import WidgetItemSearch from './WidgetItemSearch.vue'
+import WidgetItemSearch from '../item-search/WidgetItemSearch.vue'
 import WidgetSettings from '../settings/SettingsWindow.vue'
 import { AppConfig, saveConfig, pushHostConfig } from '@/web/Config'
 import LoadingAnimation from './LoadingAnimation.vue'
