+++ renderer/src/web/overlay/widgets.ts
@@ -62,8 +62,3 @@ export interface ImageStripWidget extends Widget {
     url: string
   }>
 }
-
-export interface ItemSearchWidget extends Widget {
-  anchor: Anchor
-  ocrGemsKey: string | null
-}
